530 correcting account creation errors. Also if activitites are empty do not aggregate and clear any past aggregates
