+++ app/lib/accounts.js
@@ -1,4 +1,3 @@
-const userEmailSchema = require('./userEmailSchema');
 AccountsTemplates.configure({
   // Behavior
   forbidClientAccountCreation: true,
@@ -16,42 +15,3 @@ AccountsTemplates.configureRoute('enrollAccount');
 AccountsTemplates.configureRoute('forgotPwd');
 AccountsTemplates.configureRoute('resetPwd');
 AccountsTemplates.configureRoute('signIn');
-
-Meteor.startup(function() {
-  Accounts.onCreateUser(function(options, user) {
-    let { deactivateOn } = options;
-    _.extend(user, {
-      isActive: true,
-      deactivateOn
-    });
-    return user;
-  });
-
-  Accounts.validateNewUser(function(user) {
-    const userEmailRecord = { emails: { ...user.emails } };
-    try {
-      userEmailSchema.validate(userEmailRecord);
-      return true;
-    } catch (e) {
-      // Must throw meteor error to show custom validation message
-      // Sending static error message for consistency with messages from previous validators
-      throw new Meteor.Error(500, 'Email: Invalid email');
-    }
-  });
-
-  /* Preventing login attempts by inactive users */
-  Accounts.validateLoginAttempt(function(attempt) {
-    var { isActive } = attempt.user;
-
-    /* If the user is inactive throw an error */
-    if (isActive === false) {
-      attempt.allowed = false;
-      throw new Meteor.Error(403, 'User account is inactive.');
-    }
-
-    /* else return permission set by the login method */
-    return attempt.allowed;
-  });
-
-  Accounts.emailTemplates.from = process.env.FROM_EMAIL;
-});
